s390x: fix build when long double is more than 64 bits
The floating point registers on s390x are 64-bits wide so type
them as double precision rather than long double.

Also, GCC on Ubuntu 18.04 complains about the sizeof calls so fix
those too.

+++ include/libunwind-s390x.h
@@ -51,7 +51,7 @@ extern "C" {
 typedef uint64_t unw_word_t;
 typedef int64_t unw_sword_t;
 
-typedef long double unw_tdep_fpreg_t;
+typedef double unw_tdep_fpreg_t;
 
 typedef enum
   {
